Priors

At the root of the Bayesian method are prior distributions. In ExtensibleMCMCPlots.jl these are represented by structs inheriting from

ExtensibleMCMC.Prior — Type

Supertype for all prior distributions

Each concrete type must implement a method

Distributions.logpdf — Method

logpdf(::Prior, θ)

Evaluate the log-probability density function at θ for a given prior.

Concrete types

The following two priors are important particularly for running quick tests:

ExtensibleMCMC.ImproperPrior — Type

Flat prior

ExtensibleMCMC.ImproperPosPrior — Type

Flat prior for positive coordinates

Additionally, you may leverage a full range of distributions defined in Distributions.jl by using

ExtensibleMCMC.StandardPrior — Type

struct StandardPrior{T} <: Prior
    dist::T
end

A standard prior π(θ).

ExtensibleMCMC.ProductPrior — Type

struct ProductPrior{T,K} <: Prior
    dists::T
    idx::K
end

Generic prior distribution over parameter vector θ written in a facorized form. For instance if the prior may be written as

\[π(θ) = π_1(θ_{1:k})π_2(θ_{(k+1):n}),\]

then dists would correspond to a list containing π_1 and π_2 and idx would be a list containing 1:k and (k+1):n.

ProductPrior(dists, dims)

Base consructor specifying a list of prior dsitributions dists and a corresponding number of dimensions to which each prior distribution refers to.

Warning

These last two are not properly tested, so there might be some hiccups.