Index

ExtensibleMCMC.Adaptation

ExtensibleMCMC.AdaptationUnifRW

ExtensibleMCMC.AdaptationUnifRW

ExtensibleMCMC.Callback

ExtensibleMCMC.GaussianRandomWalk

ExtensibleMCMC.GaussianRandomWalkMix

ExtensibleMCMC.GenericChainStats

ExtensibleMCMC.GenericGlobalWorkspace

ExtensibleMCMC.GenericLocalWorkspace

ExtensibleMCMC.GenericMCMCBackend

ExtensibleMCMC.GlobalWorkspace

ExtensibleMCMC.HaarioTypeAdaptation

ExtensibleMCMC.ImproperPrior

ExtensibleMCMC.LocalWorkspace

ExtensibleMCMC.MCMC

ExtensibleMCMC.MCMCBackend

ExtensibleMCMC.MCMCConjugateUpdate

ExtensibleMCMC.MCMCGradientBasedUpdate

ExtensibleMCMC.MCMCImputation

ExtensibleMCMC.MCMCParamUpdate

ExtensibleMCMC.MCMCSchedule

ExtensibleMCMC.MCMCUpdate

ExtensibleMCMC.MCMCUpdateDecorator

ExtensibleMCMC.NoAdaptation

ExtensibleMCMC.RandomWalk

ExtensibleMCMC.RandomWalkUpdate

ExtensibleMCMC.SavingCallback

ExtensibleMCMC.StandardGlobalSubworkspace

ExtensibleMCMC.StandardLocalSubworkspace

ExtensibleMCMC.TransitionKernel

ExtensibleMCMC.UniformRandomWalk

ExtensibleMCMC.Workspace

Base.:==

Base.iterate

Base.rand

DataStructures.reset!

Distributions.logpdf

ExtensibleMCMC._AdaptationUnifRW

ExtensibleMCMC._AdaptationUnifRW

ExtensibleMCMC._AdaptationUnifRW

ExtensibleMCMC.__run!

ExtensibleMCMC._adpt_rw_fill_defaults

ExtensibleMCMC._adpt_rw_fill_defaults

ExtensibleMCMC._adpt_rw_fill_defaults

ExtensibleMCMC._adpt_rw_fill_defaults

ExtensibleMCMC._assure_scalar

ExtensibleMCMC._upgrade_to_svec

ExtensibleMCMC._upgrade_to_vec

ExtensibleMCMC.accept_reject!

ExtensibleMCMC.acceptance_rate

ExtensibleMCMC.acceptance_rate!

ExtensibleMCMC.check_if_execute

ExtensibleMCMC.check_if_execute

ExtensibleMCMC.cleanup!

ExtensibleMCMC.cleanup!

ExtensibleMCMC.compute_ll!

ExtensibleMCMC.compute_δ

ExtensibleMCMC.compute_ϵ

ExtensibleMCMC.create_workspace

ExtensibleMCMC.create_workspaces

ExtensibleMCMC.custom_zero

ExtensibleMCMC.data_to_csv

ExtensibleMCMC.execute!

ExtensibleMCMC.execute!

ExtensibleMCMC.find_available_name

ExtensibleMCMC.find_starting_idx

ExtensibleMCMC.get_decorators

ExtensibleMCMC.init!

ExtensibleMCMC.init!

ExtensibleMCMC.init!

ExtensibleMCMC.init_global_workspace

ExtensibleMCMC.isdecorator

ExtensibleMCMC.isequal_except

ExtensibleMCMC.log_prior

ExtensibleMCMC.log_prior

ExtensibleMCMC.log_transition_density

ExtensibleMCMC.log_transition_density

ExtensibleMCMC.log_transition_density

ExtensibleMCMC.logit

ExtensibleMCMC.proposal!

ExtensibleMCMC.readjust!

ExtensibleMCMC.register!

ExtensibleMCMC.register_accept_reject_results!

ExtensibleMCMC.register_accept_reject_results!

ExtensibleMCMC.reschedule!

ExtensibleMCMC.run!

ExtensibleMCMC.set_proposal!

ExtensibleMCMC.sigmoid

ExtensibleMCMC.strip_decorators

ExtensibleMCMC.time_to_update

ExtensibleMCMC.transition

ExtensibleMCMC.update!

ExtensibleMCMC.update_workspaces!

ExtensibleMCMC.AdaptationUnifRW — Method

AdaptationUnifRW(θ::K; kwargs...) where K

Define an adaptation scheme for a random walker that does updates on the parameter of shape and type θ. The following named parameters can be specified ...

Arguments

adapt_every_k_steps=100: number of steps based on which adaptation happens

target_accpt_rate=0.234: acceptance rate of MH step that is to be targetted

scale=1.0: scaling of the adaptation

min=1e-12: minimum allowable (half)-range of the uniform sampler

max=1e7: maximum allowable (half)-range of the unifor sampler

offset=1e2: number of adaptation steps after which the shrinking of adaptive steps is supposed to start.

...

ExtensibleMCMC.GenericChainStats — Type

mutable struct GenericChainStats{T} <: ChainStats
    rolling_ar::Vector{Vector{Float64}}
    roll_window::Int64
    mean::Vector{T}
    cov::Matrix{T}
    N::Int64
end

Simple online statistics for the Markov chain.

ExtensibleMCMC.GenericMCMCBackend — Type

GenericMCMCBackend

A flag that no specific backend is passed.

ExtensibleMCMC.ImproperPrior — Type

ImproperPrior

Flat prior

ExtensibleMCMC.MCMC — Type

mutable struct MCMC
    updates::Vector
    updates_and_decorators::Vector
    backend::MCMCBackend
    schedule::MCMCSchedule
    workspace::GlobalWorkspace
end

A struct for defining a list of updates and specifying the backend algorithm doing the MCMC sampling. Any other functionality is done solely by the internal mechanizms of function the run!.

    MCMC(
        updt_and_decor::Vector{<:Union{MCMCUpdate,MCMCUpdateDecorator}};
        backend=GenericMCMCBackend()
    )

The main constructor of an `MCMC` struct. It accepts an array of updates and
update decorators `updates_and_decorators` which together constitute a
single MCMC step. Additionally, `backend` specialized to a particular,
overarching MCMC algorithm (if it exists). An example of a backend would be
`GuidedProposalsBackend`.

ExtensibleMCMC.get_decorators — Method

get_decorators(updt_and_decor)

Retrieve all decorators from a list

ExtensibleMCMC.isdecorator — Method

isdecorator(u)

Returns true if u is a subtype of decorators.

ExtensibleMCMC.run! — Function

run!(mcmc::MCMC, num_mcmc_steps, data, θinit, callbacks; kwargs...)

The main calling function of this package that initializes, runs and outputs the results of an MCMC sampler for discretely observed stochastic processes. mcmc specifies the sequence of updates that constitute a single MCMC step and provides additional info about the MCMC backend algorithm to be used. num_mcmc_steps is the total number of MCMC steps, data completely characterizes everything that is known about the observations (including everything known about the underlying dynamics and the way it was collected), θinit is the initial guess for the unknown parameters and callbacks is a collection of extra utility functions that do extra work around MCMC sampling. exclude_updates is a standard named argument which lists the update indices and corresponding ranges of MCMC iterations at which given updates are supposed to be omitted. Additional named arguments are passed onto initializers of global workspace and depend on the chosen MCMCBackend.

ExtensibleMCMC.Adaptation — Type

Adaptation

Supertype for all adaptation schemes.

ExtensibleMCMC.Callback — Type

Callback

Supertype of all callbacks. They specify additional actions that can be performed before or after each MCMC step or at the very end of sampling.

ExtensibleMCMC.GenericGlobalWorkspace — Type

struct GenericGlobalWorkspace{T,SW,TL} <: GlobalWorkspace{T}
    sub_ws::SW
    ll::Vector{Float64}
    P::TL
    P°::TL
end

Generic global workspace with sub_ws containing current state and keeping track of its history and some basic statistics. P and P° are the target laws with accepted and proposal state set as parameters.

ExtensibleMCMC.GenericLocalWorkspace — Type

struct GenericLocalWorkspace{T} <: LocalWorkspace{T}
    sub_ws::StandardLocalSubworkspace{T}
    sub_ws°::StandardLocalSubworkspace{T}
    acceptance_history::Vector{Bool}
end

Generic local workspace with sub_ws containing a subset of state that the corresponding updates operates on, with currently accepted value of state as well as its log-likelihood. sub_ws° corresponds to a proposal state. acceptance_history keeps track of accept/reject decisions.

ExtensibleMCMC.GlobalWorkspace — Type

GlobalWorkspace{T} <: Workspace

Supertype of all global workspaces. Each MCMC sampler must have a unique global workspace, which contains state, state_history, state_proposal_history, acceptance_history and data. state is the paramater vector that the MCMC sampling is done for (other names being self-explanatory).

ExtensibleMCMC.LocalWorkspace — Type

LocalWorkspace{T} <: Workspace

Supertype of all local workspaces. Local workspace should contain any additional gathering of objects that are needed by specific updates, but are not are not already in a global workspace. Each MCMC update has its own LocalWorkspace.

ExtensibleMCMC.MCMCBackend — Type

MCMCBackend

Supertype of all backends for the MCMC sampler.

ExtensibleMCMC.MCMCConjugateUpdate — Type

MCMCConjugateUpdate <: MCMCUpdate

Supertype of all conjugate updates, for which sampling can be done directly, without resorting to a Metropolis-Hastings algorithm

ExtensibleMCMC.MCMCGradientBasedUpdate — Type

MCMCGradientBasedUpdate <: MCMCParamUpdate

Supertype of all updates that require local gradient information to perform update of state.

ExtensibleMCMC.MCMCImputation — Type

MCMCImputation <: MCMCUpdate

Supertype of all updates that do not make any changes to the main parameter state in GlobalWorkspace, but instead, perform sampling on any auxiliary variables that are not of direct interest to the MCMC chain.

ExtensibleMCMC.MCMCParamUpdate — Type

MCMCParamUpdate <: MCMCUpdate

Supertype of all updates that make changes to the main MCMC parameter called state in the GlobalWorkspace.

ExtensibleMCMC.MCMCSchedule — Type

mutable struct MCMCSchedule
    num_mcmc_steps::Int64
    num_updates::Int64
    start::NamedTuple{(:mcmciter, :pidx),Tuple{Int64,Int64}}
    exclude_updates::DefaultDict{Int64,OrdinalRange{Int64,Int64},UnitRange{Int64}}
end

An object used for iterating over the steps of MCMC samplers. num_mcmc_steps is the total number of MCMC steps, whereas num_updates is a total number of separate updates that can be performed in a single MCMC step. Not all of those updates have to be performed and the pattern of excluding the updates is stored in exclude_updates. The num_updates and exclude_updates can change in the midst of iterating through the MCMCSchedule and the iterator will register that and act accordingly.

ExtensibleMCMC.MCMCUpdate — Type

MCMCUpdate

Supertype of all MCMC updates

ExtensibleMCMC.MCMCUpdateDecorator — Type

MCMCUpdateDecorator

Supertype of all decorators to update schemes. In this context, we use the word decorator to refer to any additional information that needs to be conveyed to the MCMC sampler that is not directly extractable from the update-objects themselves (for instance, a change in the delimitation of blocks that is made in-between updates).

ExtensibleMCMC.NoAdaptation — Type

NoAdaptation

A struct-flag for indicating that no adaptation is to be done.

ExtensibleMCMC.RandomWalk — Type

RandomWalk <: TransitionKernel

Supertype for all random walkers.

ExtensibleMCMC.SavingCallback — Type

struct SavingCallback <: Callback
    save_at_the_end::Bool
    save_intermediate::Bool
    save_at_iters::Vector{Int64}
    filename::String
end

Struct for saving the intermediate or final states of the sampled chain to the disk.

    SavingCallback(;
        save_at_the_end=true,
        save_at_iters=[],
        overwrite]=false,
        filename="mcmc_results",
        add_datestamp=false,
        path=".",
    )

The main constructor for `SavingCallback`. `save_at_the_end` is an indicator
flag for whether to save to a file at the end of mcmc sampling,
`save_at_iters` additionally specifies at which intermediate iterations
the chain should be saved to a file, `overwrite` if set to true will
overwrite any already existing file that shares the name with the one passed
to this function. `filename` is the main stem of the file's name,
`add_datestamps` will add the date and time at the time of creating the
callback to the filename and `path` specifies the directory path to save to.

ExtensibleMCMC.StandardGlobalSubworkspace — Type

struct StandardGlobalSubworkspace{T,TD} <: GlobalWorkspace{T}
    state::Vector{T}
    state_history::Vector{Vector{Vector{T}}}
    state_proposal_history::Vector{Vector{Vector{T}}}
    data::TD
    stats::GenericChainStats{T}
end

Standard containers expected to be present in every global workspace. state is the currently accepted parameter θ, state_history is a chain of states that have been accepted and state_proposal_history is a chain of states that have been proposed. data are the data passed to an MCMC sampler (usually just a pointer) and stats gathers some basic online information about the chain.

ExtensibleMCMC.StandardLocalSubworkspace — Type

struct StandardLocalSubworkspace{T} <: LocalWorkspace{T}
    state::Vector{T}
    ll::Vector{Float64}
    ll_history::Vector{Vector{Float64}}
    ∇ll::Vector{Vector{Float64}}
    momentum::Vector{Float64}
end

Standard containers likely to be present in every local workspace. state is the currently accepted subset of parameter θ that the corresponding update operates on, ll is the corresponding log-likelihood and ll_history is the chain of log-likelihoods. A single ll is of the type Vector{Float64} to reflect the fact that a problem might admit a natural factorisation into independent components that may be operated on independently, in parallel with each entry in ll corresponding to a separate subcomponent. ∇ll is the gradient of log-likelihood (needed by gradient-based algorithms) and momentum is the variable needed for the Hamiltionan dynamics. ∇ll and momentum may be simply left untouched if the problem does not need them.

ExtensibleMCMC.TransitionKernel — Type

TransitionKernel

Supertype of all transition kernels that perform updates on the main state of the global workspace. Conceptually, these are used by subtypes of MCMCParamUpdate to perform actual sampling and instances of types inheriting from TransitionKernel are usually member objects of the instances inheriting from MCMCParamUpdate.

ExtensibleMCMC.Workspace — Type

Workspace

Supertype of all workspaces–-i.e. of structs that gather in one place various objects that the MCMC sampler operates on.

Base.:== — Method

Base.:(==)(a::AdaptationUnifRW{T}, b::AdaptationUnifRW{S}) where{T,S}

Convenience comparison of AdaptationUnifRW instances. Used only for tests.

Base.iterate — Function

Base.iterate(iter::MCMCSchedule, state=iter.start)

Iterate through the MCMCSchedule, outputting the named tuples (mcmciter=..., pidx=...) along the way, indicating the current index of the mcmc sampler (mcmciter), as well as the index of update that is to be performed.

Base.rand — Method

Base.rand(rw::UniformRandomWalk, θ)

Sample a new state for a random walker rw that is currently in a state θ.

DataStructures.reset! — Method

reset!(adpt::AdaptationUnifRW)

Reset the number of proposals and accepted samples to zero.

Distributions.logpdf — Method

Distributions.logpdf(rw::UniformRandomWalk, θ, θ°)

Evaluate the log-probability density function for a move of a random walker from θ to θ°.

ExtensibleMCMC._AdaptationUnifRW — Method

_AdaptationUnifRW(
        ::Any, ::Val{:scalar}, ::Val{N}, trgt_ar, steps; kwargs...
    ) where N

Internal constructor of AdaptationUnifRW, populating entries with scalars.

ExtensibleMCMC._AdaptationUnifRW — Method

_AdaptationUnifRW(
        ::Val{false}, ::Val{:nonscalar}, v::Val{N}, trgt_ar, steps; kwargs...
    ) where N

Internal constructor of AdaptationUnifRW, populating entries with SVectors.

ExtensibleMCMC._AdaptationUnifRW — Method

_AdaptationUnifRW(
        ::Val{true}, ::Val{:nonscalar}, ::Val{N}, trgt_ar, steps; kwargs...
    ) where N

Internal constructor of AdaptationUnifRW, populating entries with vectors.

ExtensibleMCMC.__run! — Method

__run!(global_ws, local_wss, updates, schedule, callbacks)

Internal loops that run the (already initialized) MCMC sampler. global_ws and local_wss are the already initialized global workspace and local workspaces (one for each update) respectively. schedule is the iterator over MCMC steps and callbacks is a list of callbacks.

ExtensibleMCMC._adpt_rw_fill_defaults — Method

_adpt_rw_fill_defaults(N; kwargs...)

Return a vector with set vector parameters, falling back on default values if necessary.

ExtensibleMCMC._adpt_rw_fill_defaults — Method

_adpt_rw_fill_defaults(f::Function)

Return a vector with set parameters, setting them according to the "recipe" f, falling back on default values if necessary.

ExtensibleMCMC._adpt_rw_fill_defaults — Method

_adpt_rw_fill_defaults(; kwargs...)

Return a vector with set scalar parameters, falling back on default values if necessary.

ExtensibleMCMC._adpt_rw_fill_defaults — Method

_adpt_rw_fill_defaults(N; kwargs...)

Return a vector with set SVector parameters, falling back on default values if necessary.

ExtensibleMCMC._assure_scalar — Method

_assure_scalar(v)

Accept a scalar or a vector of length one and return it as a scalar. Raise assertion error if other objects are passed.

ExtensibleMCMC._upgrade_to_svec — Method

_upgrade_to_svec(v, ::Val{N}) where N

Receive a scalar or a vector of length 1 and return a length N static vector of repeats of entry v.

ExtensibleMCMC._upgrade_to_vec — Method

_upgrade_to_vec(v, N)

Receive a scalar or a vector of length 1 and return a length N vector of repeats of entry v.

ExtensibleMCMC.accept_reject! — Function

accept_reject!(updt, global_ws, ws, step, i=1)

Finish computations of the log-likelihood ratio between the target and proposal in MH acceptance probability and accept/reject respectively.

ExtensibleMCMC.acceptance_rate! — Method

acceptance_rate!(adpt::AdaptationUnifRW)

Destructive computation of a current acceptance rate that also resets the number of proposals and accepted samples to zeros.

ExtensibleMCMC.acceptance_rate — Method

acceptance_rate(adpt::AdaptationUnifRW)

Compute current acceptance rate of the Metropolis-Hastings update step

ExtensibleMCMC.check_if_execute — Method

check_if_execute(callback::Callback, step, flag)

Check if callback is supposed to be executed at the mcmc step step with pre- post- update flag flag.

ExtensibleMCMC.check_if_execute — Method

check_if_execute(callback::SavingCallback, iter)

Return true if the mcmc iteration iter is one at which an intermediate save is to be made.

ExtensibleMCMC.cleanup! — Method

cleanup!(callback::Callback, ws::GlobalWorkspace, step)

The last call to callback, after all MCMC steps and before exiting the function run.

ExtensibleMCMC.cleanup! — Method

cleanup!(callback::SavingCallback, ws::GlobalWorkspace, iter)

Save the entire MCMC chain, history of proposals, acceptance history etc.

ExtensibleMCMC.compute_ll! — Method

compute_ll!(
    global_ws::GlobalWorkspace,
    ws::LocalWorkspace,
    step,
)

Evaluate the proposal log-likelihood at the observations.

ExtensibleMCMC.compute_δ — Method

compute_δ(p, mcmc_iter)

δ decreases roughly proportional to scale/sqrt(iteration)

ExtensibleMCMC.compute_ϵ — Function

compute_ϵ(ϵ_old, p, a_r, δ, flip=1.0, f=identity, finv=identity)

ϵ is moved by δ to adapt to target acceptance rate

ExtensibleMCMC.create_workspace — Method

create_workspace(
    ::GenericMCMCBackend,
    mcmcupdate,
    global_ws::GenericGlobalWorkspace{T},
    num_mcmc_steps
) where {T}

Create a local workspace for a given mcmcupdate.

ExtensibleMCMC.create_workspaces — Method

create_workspaces(v::GenericMCMCBackend, mcmc::MCMC)

Create local workspaces, one for each update.

ExtensibleMCMC.custom_zero — Method

custom_zero(x::T, ::Type{elT}) where {T,elT}

Create a zero with eltype elT, that is of the same structure as the collection x.

Examples

julia> custom_zero(3.0, Bool)
false
julia> custom_zero([3.0, 4.0], Bool)
2-element Array{Bool,1}:
 0
 0
julia> custom_zero(SVector{2}(1, 2), ComplexF64)
2-element SArray{Tuple{2},Complex{Float64},1,2} with indices SOneTo(2):
 0.0 + 0.0im
 0.0 + 0.0im

ExtensibleMCMC.data_to_csv — Method

data_to_csv(f, ws::GlobalWorkspace, i)

Write data entries to a CSV file.

ExtensibleMCMC.execute! — Method

execute!(
    callback::Callback,
    global_ws::GlobalWorkspace,
    local_ws::LocalWorkspace,
    step,
    flag
)

Execute the callback.

ExtensibleMCMC.execute! — Method

execute!(sc::SavingCallback, ws::GlobalWorkspace, iter)

Save the chain of accepted states, proposed states and acceptance history to the disk.

ExtensibleMCMC.find_available_name — Function

find_available_name(path, filename, disambig_num, extension=".csv")

Check if a chosen filename already exists, if so, then tries appending consecutive numbers to the end of the file until the first one that is not used yet is found.

ExtensibleMCMC.find_starting_idx — Method

find_starting_idx(callback::SavingCallback, iter)

Find the last index for which saving was done.

ExtensibleMCMC.init! — Function

init!(
    mcmc::MCMC,
    num_mcmc_steps,
    data,
    θinit,
    exclude_updates=[];
    kwargs...
)

Initialize the schedule and the global workspace of the MCMC sampler for a given set of updates (already saved in mcmc.updates), a total number of MCMC iterations given by num_mcmc_steps, observed dataset data. θinit is the initial value of the main parameter that the MCMC sampling is done for, exclude_updates lists the update indices and the repspective ranges of mcmc iterations from which these updates are supposed to be omitted from and kwargs lists all additional named arguments passed for creating a global workspace.

ExtensibleMCMC.init! — Method

init!(::Callback)

Initialization actions for a callback. By default nothing to be done.

ExtensibleMCMC.init! — Method

init!(callback::SavingCallback)

Create a CSV file for writing into.

ExtensibleMCMC.init_global_workspace — Method

init_global_workspace(
    ::GenericMCMCBackend,
    schedule::MCMCSchedule,
    updates::Vector{<:MCMCUpdate},
    data,
    θinit::Vector{T};
    kwargs...
) where T

Initialize the GenericGlobalWorkspace.

ExtensibleMCMC.isequal_except — Method

isequal_except(
    ::AdaptationUnifRW{T}, b::AdaptationUnifRW{S}, args...
    ) where{T,S}

Convenience comparison of AdaptationUnifRW instances. Used only for tests. The fields specified by args are excluded from comparison.

ExtensibleMCMC.log_prior — Method

log_prior(updt::MCMCParamUpdate, ws::LocalWorkspace)

Evaluate the log-prior at θ.

ExtensibleMCMC.log_prior — Method

log_prior(updt::MCMCParamUpdate, ws::LocalWorkspace, ::Proposal)

Evaluate the log-prior at θ°.

ExtensibleMCMC.log_transition_density — Method

log_transition_density(::Previous, updt::MCMCParamUpdate, ws::LocalWorkspace)

Evaluate the log-density for a transition θ → θ°.

ExtensibleMCMC.log_transition_density — Method

log_transition_density(
    ::Proposal,
    updt::MCMCParamUpdate,
    ws::LocalWorkspace
)

Evaluate the log-density for a transition θ° → θ.

ExtensibleMCMC.logit — Function

logit(x, a=1.0)

Logit function. (Inverse of sigmoid).

ExtensibleMCMC.proposal! — Method

proposal!(updt::RandomWalkUpdate, global_ws, ws::LocalWorkspace, step)

Sample a proposal value for the MCMC's local state variable.

ExtensibleMCMC.readjust! — Method

readjust!(rw::UniformRandomWalk, adpt::AdaptationUnifRW, mcmc_iter)

Adaptive readjustment of the range for sampling uniforms by the random walker.

ExtensibleMCMC.register! — Method

register!(adpt::AdaptationUnifRW, accepted::Bool, ::Any)

Register the result of the acceptance decision in the Metropolis-Hastings step.

ExtensibleMCMC.register_accept_reject_results! — Function

register_accept_reject_results!(
    accepted::Bool, updt, ws::LocalWorkspace, step, i=1
)

Register the results of accept/reject step relevant to a local workspace.

ExtensibleMCMC.register_accept_reject_results! — Function

register_accept_reject_results!(
    accepted::Bool,
    updt,
    global_ws::GlobalWorkspace,
    local_ws::LocalWorkspace,
    step,
    i=1
)

Register the result of accept/reject step.

ExtensibleMCMC.reschedule! — Function

reschedule!(
    schedule::MCMCSchedule,
    num_new_updates=0,
    idxes_to_remove=[],
    idxes_to_add=[]
)

Make changes to the schedule (possibly in the midst of iterating through it). Add num_new_updtes-many allowable updates that can be performed at each MCMC step, idxes_to_remove lists all parameter udpates that are to be completely removed from the MCMC sampler, idxes_to_add lists entries that are to be added to schedule.exclude_updates.

ExtensibleMCMC.set_proposal! — Method

set_proposal!(
    global_ws::GlobalWorkspace,
    ws::LocalWorkspace,
    loc2glob,
    step
)

The proposal parameter θ° has been sampled by this point. This function propagates it through workspaces and sets the new parameter in proposal law P°

ExtensibleMCMC.sigmoid — Function

sigmoid(x, a=1.0)

Sigmoid function. (Inverse of logit).

ExtensibleMCMC.strip_decorators — Method

strip_decorators(updt_and_decor)

Remove all decorators from a list

ExtensibleMCMC.time_to_update — Method

time_to_update(adpt::AdaptationUnifRW)

Return true if it's the time to update the ϵ parameter of Uniform random walks

ExtensibleMCMC.transition — Method

transition(schedule::MCMCSchedule, state)

Determine the next state of the MCMCSchedule iterator, by skipping through all updates that are to be excluded, as per schedule.exclude_updates.

ExtensibleMCMC.update! — Method

update!(
    updt::MCMCParamUpdate,
    global_ws::GlobalWorkspace,
    ws::LocalWorkspace,
    step,
)

General recipe for performing parameter update (doing an update of the main state of the MCMC sampler).

ExtensibleMCMC.update_workspaces! — Method

update_workspaces!(
    local_updt::MCMCParamUpdate,
    global_ws::GlobalWorkspace,
    local_ws::LocalWorkspace,
    step,
)

Transfer all the information that is needed at the time of starting the MCMC update step from the global to the local workspace.